refactor(bundler): рефакторинг
