+++ src/interfaces.ts
@@ -7,30 +7,23 @@ export interface ITableClass {
     elListNums: HTMLUListElement;
     elWrapCells: HTMLDivElement;
     _countNums: number;
+    _startX: number|null;
+    _currentRowWidth: number|null;
+    _currentRow: HTMLDivElement|null;
 
     _fillNums(): Array<number>;
     _fillCells(): Array<ICell>;
     _fillLetters(lettersArr?: Array<string>, iteration?: number): Array<string>;
     _getLengthEnglishAlphabet(): number;
-    _resize(): void;
+    _resizeHandler(): void;
+    _startResizeRow(e: MouseEvent, thin: HTMLSpanElement): void;
+    _stopResizeRow(): void;
+    _mouseResizeRow(e: MouseEvent): void;
+    _initEventsToResizeRow(): void;
     renderCells(): void;
     renderNums(): void;
     clearNums(): void;
     render(): void;
-}
-
-export interface ILetterClass {
-    _elLetters: NodeListOf<HTMLDListElement>;
-    _elCells: NodeListOf<HTMLLIElement>;
-    _elLetterThins: NodeListOf<HTMLSpanElement>;
-    _startX: number|null;
-    _currentRowWidth: number|null;
-    _currentRow: HTMLDivElement|null;
-
-    _stopResizeRow(): void;
-    _mouseResizeRow(e: MouseEvent): void;
-    _startResizeRow(e: MouseEvent, thin: HTMLSpanElement): void;
-    init(): void;
 }
 
 export interface ICellClass {
